Refactor Android project structure and update package names
