+++ android/app/src/main/AndroidManifest.xml
@@ -28,6 +28,9 @@
         android:name="android.permission.POST_NOTIFICATIONS"
         android:targetSdkVersion="33" />
     <uses-permission android:name="android.permission.READ_PHONE_STATE"/>
+    
+    <!-- Google Play Services permission for Advertising ID (required for Android 13+) -->
+    <uses-permission android:name="com.google.android.gms.permission.AD_ID"/>
 
     <uses-feature android:name="android.hardware.camera" />
     <uses-feature android:name="android.hardware.camera.autofocus" android:required="false" />
