--- android/app/src/main/java/com/galaxyrn/MainActivity.java
@@ -1,97 +1,0 @@
-package com.galaxyrn;
-
-import android.content.Intent;
-import android.media.AudioManager;
-import android.os.Bundle;
-import android.util.Log;
-import android.os.Process;
-import android.content.IntentFilter;
-import android.content.Context;
-import android.view.WindowManager;
-
-import androidx.annotation.NonNull;
-
-import com.facebook.react.ReactActivity;
-import com.facebook.react.ReactActivityDelegate;
-import com.facebook.react.bridge.ReactContext;
-import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.defaults.DefaultReactActivityDelegate;
-import com.facebook.react.modules.core.DeviceEventManagerModule;
-import com.galaxyrn.foreground.ForegroundService;
-import com.galaxyrn.permissions.PermissionHelper;
-import com.facebook.react.ReactInstanceManager;
-import com.galaxyrn.logger.GxyLogger;
-import com.galaxyrn.logger.GxyLoggerUtils;
-
-public class MainActivity extends ReactActivity {
-    private static final String TAG = "MainActivity";
-    private PermissionHelper permissionHelper;
-
-    /**
-     * Returns the name of the main component registered from JavaScript.
-     * This is used to schedule rendering of the component.
-     */
-    @Override
-    protected String getMainComponentName() {
-        return "GalaxyRN";
-    }
-
-    @Override
-    protected ReactActivityDelegate createReactActivityDelegate() {
-        return new DefaultReactActivityDelegate(this, getMainComponentName(), false);
-    }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setVolumeControlStream(AudioManager.STREAM_VOICE_CALL);
-
-        permissionHelper = new PermissionHelper(this);
-
-        // Using custom logger instead of Log.d
-        GxyLogger.i(TAG, "onCreate");
-        GxyLoggerUtils.logDeviceInfo(TAG);
-
-        getReactInstanceManager().addReactInstanceEventListener(new ReactInstanceManager.ReactInstanceEventListener() {
-            @Override
-            public void onReactContextInitialized(ReactContext context) {
-                GxyLogger.i("ReactContext",
-                        "Updating PermissionHelper with ReactApplicationContext. Permissions ready: "
-                                + permissionHelper.permissionsReady);
-                if (!permissionHelper.permissionsReady) {
-                    permissionHelper.initModules((ReactApplicationContext) context);
-                } else {
-                    permissionHelper.sendPermissions();
-                }
-            }
-        });
-    }
-
-    @Override
-    protected void onDestroy() {
-        GxyLogger.d(TAG, "onDestroy - ensuring all services are stopped");
-        super.onDestroy();
-
-        try {
-            setVolumeControlStream(AudioManager.USE_DEFAULT_STREAM_TYPE);
-            GxyLogger.i(TAG, "Activity cleanup completed");
-        } catch (Exception e) {
-            GxyLogger.e(TAG, "Error during activity cleanup", e);
-        }
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
-            @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        permissionHelper.handlePermissionResult(requestCode, permissions, grantResults);
-    }
-
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if (permissionHelper != null) {
-            permissionHelper.onActivityResult(requestCode, resultCode, data);
-        }
-    }
-}
